functions/requirements.txt
azure-functions==1.8.0
azure-servicebus==7.5.0
redis==4.1.2